backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
openai==1.97.0
python-dotenv==1.0.1
pydantic==2.9.2
groq==0.12.0
requests==2.32.3
